DEBUG ON!

+++ workbench/cw/crowdflower/src/Cw/Crowdflower/crowdflower.php
@@ -126,7 +126,7 @@ class Crowdflower extends \FrameWork {
 				//  - Tags / keywords
 				//  - Worker levels (defaults to '1')
 				//  - Expiration?
-				$debug = false;
+				$debug = true;//false;
 
 				if($debug) {
 					print "\r\n\r\nRESULT";
@@ -134,7 +134,7 @@ class Crowdflower extends \FrameWork {
 				}	
 
 				$csvresult = $this->CFJob->uploadInputFile($id, $csv);
-				unlink($csv); // DELETE temporary CSV.
+				if(!$debug) unlink($csv); // DELETE temporary CSV.
 
 				if($debug) {
 					print "\r\n\r\nCSVRESULT";
